reformulation for solving S only
Changes to solve hs1D by solving only water stored.
We stays with ode15s solver as it is adapted to stiff problems.
To detect changes in precipitations, we strongly advise to chose a MaxStep in the ode15s options (eg 3600s for an hourly precipitation record).
Q and QS are computed in a postprocessing step.
Chnages in boussinesq_simulation essentially