Update transport.m to take into account spatialized recharge events
Modify the size of N properties in transport.m -> Now it can be a 2D matrix to encounter for the existence of spatialized recharge

Change the location where the transpose of traj_temp is computed. Check if this does not impact the computation

Modification so that N can be in 2D. All methods have been modified accordingly. To check on a non regression test.