Update transport.m to take into account ET
if ET is negative then the particle is retrieved from the porous medium (aquifer or soil compartment), and the particle is not further considered (this has to be checked) in the TTDs/RTDs computation.

Would be better to let it in the medium but to "tag" it not to consider it only in the TTD computation.