Update transport_2D.m
Update transport_2D by concatenating two methods compute_trajectories and cut_saturated_areas in one method compute_trajectories2

This method has been checked in steady state compared to transport.m methods. Results are in accordance with previous object with a better representation of the stratification of the flow (due to the explicit dependance on z).