Update transport_2D_par_temp.m
choice possible to comment some line in compute_trajectories method to speed up the code. The code in this case only retains the time in and the time out and so the transit time. Whereas the full option enables to derive the position of the flow path and the location of the particle during the tracking.