Update ttds.m
In the method compute_ttds : update the pdf formula. Now based on de Marsily (1986) Quantitative hydrogeology book.

Add also a gamma distribution for the particles that directly falls on saturated areas. The transit times depends on their distance to the river, but is always lower than the lowest one in the groundwater.