Update transport_2D_par_temp.m
Modifications to take into account the possibility for the particles to exit as ET. Add also a tagging to track the different flow paths in the particles transit times (namely DGW, RF, DPSA and ET).
#WOULD NEED TO CLEAN THE CODE NOW

Changes prior to this and not committed before regards the mat_pos_allocated_x_z matrix. Mauybe this is not operationnal right now.
## CHECK THIS